Use Python 3.11
Replace the use of the `amazonlinux:2` Docker image that is locked to
Python 3.7 with the official Python Docker image instead.

Choose the latest version of Python that AWS Lambda currently supports.

The offical Python images are based on Debian so change
RedHat/AmazonLinux commands to Debian equivalents.

Download and extract the Debian version of ClamAV.

Wehn zipping up the Python site-packages I excluded some files and
directories to have the ZIP equivalent to what was produced under
AmazonLinux.

Also have the pipeline confirm the Docker build works in non-master and
non-test-* branches.