PK-847 More fixes
The code executed `ld --verbose` and parse it for `SEARCH_DIR("=$PATH")`
statements to build up an `LD_LIBRARY_PATH` and then it would append
`:$CLAMAVLIB_PATH` to the end of that when calling the freshclam binary.
However the `ld` binary is not found.
It used to be available as `/usr/bin/ld` in the Python 3.7 Lamba
execution environments but is no longer for Python 3.9.
The runtimes already have `LD_LIBRARY_PATH` set, so remove all the code
that tries to build it up and just append `:$CLAMAVLIB_PATH` to the
existing value.
This is all a hack though and I've added a note to `README.md` for the
correct approach that should be taken if we want to diverge more from
upstream.

Remove the test in `clamav_test.py` that checks that the call to `ld`
works since it is no longer relevant.

Unlock the `certifi` package since that has the CA certificate bundle
which should be kept up to date.

Install the Python modules into a directory of our own making instead of
having them go into the system's
`/usr/local/lib/python${VER}/site-packages` directory.
That means you don't have to pass a stupidly long `--exclude` parameter
to the `zip` command when packaging the Python modules into the ZIP
file.

Add a `Healthengine notes` section to the `README.md` file.