Rebase against upstream/master
The upstream has updated things to Python 3.12.
Therefore rebase this to match their updates.

As part of that, ditch all our custom Docker changes and just use the
Dockerfile that upstream is using except the FROM statements since they
use a private repository that I think is a mirror.

Update the Buildkite plugins.

Tweak our README.md changes to match the new state of play.

The only code differences between us and upstream is in scan.py where
we introduced an `S3-BATCH` event source.