Update to Python 3.12
Update the Docker image to be based on Python 3.12.

Upgrade the version of ClamAV to be current.

Strip the libclamav_rust.a file since it is massive and makes the ZIP
file too large for Lambda to install otherwise.
It shrinks it from 236MB to 20MB.

Update the README.md appropriately.

Update Buildkite plugins.

Incorporate some of the changes from upstream to allow us to get to
Python 3.12.
- Switch from nosetest to pytest changing assertEquals to assertEqual.
- Replace `pytz.utc` with `datetime.UTC`.
- Create `strtobool` function replacing `distutils.util.strtobool`
- Removed `metrics.py`, the use of it, and the `datadog` import.
- Unpin versions of modules in the requirements files.